Add files via upload
Room - The default constructor for room has four doors instead of two now.
Dungeon - Room generating algorithm has been refined. Making the distribution of the number of doors off each room to be relative to dungeon difficultly could be implemented in the future with a bit of testing.
DungeonTest - Added the room location of the first room to the printed array. Added the ability to print the number of rooms generated.