Add files via upload
Bot - Added a method to change the location of the bot based on an input direction.
Room - Added a boolean that flags whether a bot is to be generated in the room or not. A bot is created in a random location in the room for now. Methods have been added to allow the bot to move randomly one unit at a time without hitting walls or walking through doors.
RoomTest - Added simple tests to see where the bot is moving.
Dungeon - The rooms generated in this class now take the botFlag boolean as a parameter. It is set to "true" by default at the moment.
Location - Minor formatting changes.