Add files via upload
Stopped player dying upon spawn with a step counter. Step counter is shown on the screen and could potentially act as a score in the game. The bot moving randomly does add a strategic element to the game which makes minimising the number of steps taken to collect the necessary gold challenging.