Add files via upload
Room - made the gold spawned in larger rooms smaller. Gold amount will be more random in larger rooms also.
Dungeon - the distribution on the number of doors generated is a little tighter. It will still generate quite a lot of rooms occasionally but that if difficult to control because of the nature of recursion. Since the room range is more appropriate, the numberOfRooms parameter for new dungeons in map could be made much larger to ensure there are no doors leading nowhere.