Add files via upload
Map - v1.8 Added a win condition so that the game ends when the players steps on the exitLocation after the gold requirement is met on the final level. One gold per level and only 2 levels for testing purposes.
Room - Minor changes to gold amount.
Dungeon - No stray doors leading to nowhere when limiting the number of rooms in the dungeon now.